Milestone/1.13

+++ app/src/main/java/com/ruuvi/station/tagsettings/domain/CsvExporter.kt
@@ -97,7 +97,7 @@ class CsvExporter(
                     fileWriter.append(',')
                     fileWriter.append(reading.measurementSequenceNumber.toString())
                     fileWriter.append(',')
-                    fileWriter.append(reading.txPower.toString())
+                    fileWriter.append(reading.txPower.toInt().toString())
                 }
                 fileWriter.append('\n')
             }
